exception handling moved to controller

+++ src/main/java/com/adventure/base/controller/AbstractUserController.java
@@ -2,30 +2,49 @@ package com.adventure.base.controller;
 
 import com.adventure.base.model.User;
 import com.adventure.base.service.UserService;
+import com.adventure.base.util.exception.UserNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+
+import java.util.Optional;
 
 public abstract class AbstractUserController {
 
     protected UserService userService;
 
     @Autowired
-    public AbstractUserController(UserService userService){
+    public AbstractUserController(UserService userService) {
         this.userService = userService;
     }
 
     public ResponseEntity<User> getOne(int id) {
+
+        Optional<User> user = userService.getOneById(id);
+
+        if (user.isEmpty()) {
+            throw new UserNotFoundException(String.valueOf(id));
+        }
+
         return ResponseEntity.ok().body(
-                userService.getOneById(id));
+                user.get());
     }
 
     public void delete(int id) {
-        userService.delete(id);
-        ResponseEntity.ok().build();
+        if (userService.checkExistence(id)) {
+            userService.delete(id);
+        } else {
+            throw new UserNotFoundException(String.valueOf(id));
+        }
     }
 
-    public ResponseEntity<User> getAnother(String name){
-        return ResponseEntity.ok(
-                userService.getByName(name));
+    public ResponseEntity<User> getAnother(String name) {
+
+        Optional<User> user = userService.getByName(name);
+
+        if (user.isEmpty()) {
+            throw new UserNotFoundException(name);
+        }
+
+        return ResponseEntity.ok(user.get());
     }
 }
